Load pcre-4.2 into code/trunk.
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@67 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ internal.h
@@ -598,7 +598,7 @@ typedef struct recursion_info {
 doing the matching, so that they are thread-safe. */
 
 typedef struct match_data {
-  int    match_call_count;      /* As it says */
+  unsigned long int match_call_count; /* As it says */
   unsigned long int match_limit;/* As it says */
   int   *offset_vector;         /* Offset vector */
   int    offset_end;            /* One past the end */
